split code into new components
