+++ src/Display/DisplayLogic.js
@@ -1,154 +1,4 @@
-import theme from '../theme';
-import validateSchema from '../validateSchema';
 import { getMaxSteps } from '../lib';
-import { getHeaders, getRows, getCardTable } from './displayTableMarkup';
-
-const getTruthValFromCoordinates = (numRows, row, col) => 0 === Math.floor(row / (numRows / Math.pow(2, col + 1)) % 2);
-
-const getTableDimensions = (numSentenceLetters) => {
-  return {
-    numCols: numSentenceLetters, 
-    numRows: Math.pow(2, numSentenceLetters)
-  };
-};
-
-const getLegend = (sentenceLetters) => {
-  const { numRows, numCols } = getTableDimensions(sentenceLetters?.length);
-  const legend = {};
-  for (let col = 0; col < numCols; col++) {
-    const { value: legendKey }= sentenceLetters[col];
-    legend[legendKey] = [];
-    for (let row = 0; row < numRows; row++) {
-      legend[legendKey].push(getTruthValFromCoordinates(numRows, row, col));
-    }
-  }
-  return legend;
-};
-
-const getTableModel = (schema, numRows, legend) => {
-  const tableModel = [];
-  for (let row = 0; row < numRows; row++) {
-    const rowData = [];
-    schema.forEach(el => {
-      const { elType, value } = el;
-      let cellValue;
-      if (elType === 'L' && legend[value]) {
-        cellValue = legend[value][row];
-      }
-      if (elType === 'G') {
-        cellValue = value;
-      }
-      rowData.push(cellValue);
-    });
-    tableModel.push(rowData);
-  }
-  return tableModel;
-};
-
-
-
-const setupOpMap = (schema) => {
-  const map = {};
-  const opMap = {};
-  schema.forEach((sym, i) => {
-    if (sym.elType === 'O' || sym.elType === 'N') {
-      const { value } = sym;
-      if (map[value]) {
-        map[value]++;
-      } 
-      else {
-        map[value] = 1;
-      } 
-      const id = `${value}${map[value]}`; // operator ids
-      sym.opMapId = id;
-      opMap[id] = i;
-    }
-  });
-  return opMap;
-};
-
-const getDeepestScopes = (schema) => {
-  let indeces = [];
-  let depth = 0, max = 0, start, end;
-  schema.forEach((el, i) => {
-    if (el?.value === '(') {
-      depth++;
-      if (depth > max) {
-        max = depth;
-      }
-      if (depth >= max) {
-        start = i;
-      }
-    }
-    if (el?.value ===')') {
-      if (depth === max) {
-        end = i;
-        indeces.push([start, end]);
-      }
-      depth--;
-    }
-  });
-  return indeces;
-};
-
-const getNextOperator = (schema) => {
-  let index, innermostNot;
-  for (let prec = 0; prec <= 4; prec++) {
-    for (let i = 0; i < schema.length; i++) {
-      if (schema[i]?.precedence === prec) {
-        index = i;
-        if (prec === 0) {
-          innermostNot = index;
-        }
-        else {
-          return innermostNot >= 0 ? innermostNot : index;
-        }
-      } 
-    }
-  }
-  return index;
-};
-
-const computeOperator = (data) => {
-  const { schema, opIndex, table, numRows } = data;
-  const { value: operator, opMapId } = schema[opIndex];
-  let result = [],  numArgs = 2;
-  for (let row = 0; row < numRows; row++) {
-    let value;
-    let L, R;
-    if (opIndex - 1 >= 0) L = table[row][opIndex - 1];
-    if (opIndex + 1 < table[row].length) R = table[row][opIndex + 1];
-    switch (operator) {
-      case '\u00AC': // not
-        value = !R; 
-        numArgs = 1;
-        break;
-      case '\u2228': // or
-        value = L || R; 
-        break;
-      case '\u2227': // and
-        value = L && R;
-        break;
-      case '\u21D2': // if
-        value = !L || R; 
-        break;
-      case '\u21D4': // iff
-        value = L === R;
-        break;
-      case '\u22BB': // xor
-        value = L !== R;
-        break;
-      default:
-    }
-    result.push(value);
-  }
-  const resultObj = {
-    result: result,
-    numArgs: numArgs,
-    opMapId: opMapId
-  }
-  return resultObj;
-};
 
 const computeTable = (tableData) => {
   const { schema, tableModel, numRows } = tableData;
@@ -199,7 +49,183 @@ const computeTable = (tableData) => {
   });
   if (mainOpIndex >= 0) scopedIndex = mainOpIndex;
   return scopedIndex;
-};
+}
+
+const getTableDimensions = (numSentenceLetters) => {
+  return {
+    numCols: numSentenceLetters, 
+    numRows: Math.pow(2, numSentenceLetters)
+  };
+}
+
+const getLegend = (sentenceLetters) => {
+  const { numRows, numCols } = getTableDimensions(sentenceLetters?.length);
+  const legend = {};
+  for (let col = 0; col < numCols; col++) {
+    const { value: legendKey }= sentenceLetters[col];
+    legend[legendKey] = [];
+    for (let row = 0; row < numRows; row++) {
+      legend[legendKey].push(getTruthValFromCoordinates(numRows, row, col));
+    }
+  }
+  return legend;
+}
+
+const getTableModel = (schema, numRows, legend) => {
+  const tableModel = [];
+  for (let row = 0; row < numRows; row++) {
+    const rowData = [];
+    schema.forEach(el => {
+      const { elType, value } = el;
+      let cellValue;
+      if (elType === 'L' && legend[value]) {
+        cellValue = legend[value][row];
+      }
+      if (elType === 'G') {
+        cellValue = value;
+      }
+      rowData.push(cellValue);
+    });
+    tableModel.push(rowData);
+  }
+  return tableModel;
+}
+
+const getTableButtonHandlers = (stateObj) => {
+  const { data, setData, setSchema } = stateObj;
+  const maxSteps = getMaxSteps(data?.symbols);
+  const onCollapse = () => setData({...data, collapsed: !data.collapsed});
+  const onPrev = () => {
+    if (data.steps > 0)  {
+      data.steps--;
+      setData({...data});
+    }
+  };
+  const onNext = () => {
+    if (data.steps < maxSteps) {
+      data.steps++;
+      setData({...data});
+    }
+  };
+  const onEdit = () => {
+    setData(null);
+    setSchema({...data, steps: maxSteps});
+  };
+  const onDelete = () => setData(null);
+
+  return {
+    onEdit: onEdit,
+    onDelete: onDelete,
+    onNext: onNext,
+    onPrev: onPrev,
+    onCollapse: onCollapse,
+  }
+}
+
+const getTruthValFromCoordinates = (numRows, row, col) => 0 === Math.floor(row / (numRows / Math.pow(2, col + 1)) % 2);
+
+const setupOpMap = (schema) => {
+  const map = {};
+  const opMap = {};
+  schema.forEach((sym, i) => {
+    if (sym.elType === 'O' || sym.elType === 'N') {
+      const { value } = sym;
+      if (map[value]) {
+        map[value]++;
+      } 
+      else {
+        map[value] = 1;
+      } 
+      const id = `${value}${map[value]}`; // operator ids
+      sym.opMapId = id;
+      opMap[id] = i;
+    }
+  });
+  return opMap;
+}
+
+const getDeepestScopes = (schema) => {
+  let indeces = [];
+  let depth = 0, max = 0, start, end;
+  schema.forEach((el, i) => {
+    if (el?.value === '(') {
+      depth++;
+      if (depth > max) {
+        max = depth;
+      }
+      if (depth >= max) {
+        start = i;
+      }
+    }
+    if (el?.value ===')') {
+      if (depth === max) {
+        end = i;
+        indeces.push([start, end]);
+      }
+      depth--;
+    }
+  });
+  return indeces;
+}
+
+const getNextOperator = (schema) => {
+  let index, innermostNot;
+  for (let prec = 0; prec <= 4; prec++) {
+    for (let i = 0; i < schema.length; i++) {
+      if (schema[i]?.precedence === prec) {
+        index = i;
+        if (prec === 0) {
+          innermostNot = index;
+        }
+        else {
+          return innermostNot >= 0 ? innermostNot : index;
+        }
+      } 
+    }
+  }
+  return index;
+}
+
+const computeOperator = (data) => {
+  const { schema, opIndex, table, numRows } = data;
+  const { value: operator, opMapId } = schema[opIndex];
+  let result = [],  numArgs = 2;
+  for (let row = 0; row < numRows; row++) {
+    let value;
+    let L, R;
+    if (opIndex - 1 >= 0) L = table[row][opIndex - 1];
+    if (opIndex + 1 < table[row].length) R = table[row][opIndex + 1];
+    switch (operator) {
+      case '\u00AC': // not
+        value = !R; 
+        numArgs = 1;
+        break;
+      case '\u2228': // or
+        value = L || R; 
+        break;
+      case '\u2227': // and
+        value = L && R;
+        break;
+      case '\u21D2': // if
+        value = !L || R; 
+        break;
+      case '\u21D4': // iff
+        value = L === R;
+        break;
+      case '\u22BB': // xor
+        value = L !== R;
+        break;
+      default:
+    }
+    result.push(value);
+  }
+  const resultObj = {
+    result: result,
+    numArgs: numArgs,
+    opMapId: opMapId
+  }
+  return resultObj;
+}
 
 const doOperations = (schemaData) => {
   const { subSchema, subTable, tableModel, opMap, numRows } = schemaData;
@@ -233,88 +259,12 @@ const doOperations = (schemaData) => {
     stepsLeft: steps
   }
   return finalResult;
-};
-
-const getSchemaTable = (tableData) => {
-  const { schema, sentenceLetters, clickHandlers, showButtons } = tableData;
-  let { symbols, steps } = schema;
-  const { numRows } = getTableDimensions(sentenceLetters.length);
-  const legend = getLegend(sentenceLetters);
-  const tableModel = getTableModel(symbols, numRows, legend);
-  let schemaTableHeaders = getHeaders(symbols);
-  let schemaTable = getRows(tableModel);
-  const style = { 
-    display: symbols.length === 0 && 'none',
-    backgroundColor: !showButtons ? theme.palette.primary.dark : theme.palette.grey['700']
-  }
-
-  if (validateSchema(symbols)) {
-    const mainOpIndex = computeTable({
-      tableModel: tableModel,
-      schema: schema, 
-      numRows: numRows,
-      steps: steps
-    });
-    schemaTableHeaders = getHeaders(symbols, mainOpIndex);
-    schemaTable = getRows(tableModel, mainOpIndex);
-  }
-  return getCardTable({
-    style: style,
-    headers: schemaTableHeaders,
-    table: schemaTable,
-    clickHandlers: clickHandlers,
-    showButtons: showButtons,
-    collapsed: schema.collapsed,
-    schemaType: schema.type,
-  });
-};
-
-const getSavedPremiseTables = (tableData) => {
-  let { premises } = tableData;
-  const { sentenceLetters, setPremises, setSchema } = tableData;
-  return premises?.map((premise, i) => getSchemaTable({
-    sentenceLetters: sentenceLetters,
-    schema: {...premise},
-    showButtons: true,
-    clickHandlers: getTableButtonHandlers({
-      data: premise,
-      setData: (data) => {
-        data ? premises[i] = {...data} : premises.splice(i, i + 1); // delete premise when data is null, update otherwise
-        setPremises([...premises]);
-      },
-      setSchema: setSchema
-    })
-  }));
-};
-
-const getTableButtonHandlers = (stateObj) => {
-  const { data, setData, setSchema } = stateObj;
-  const maxSteps = getMaxSteps(data?.symbols);
-  const onEdit = () => {
-    setData(null);
-    setSchema({...data, steps: maxSteps});
-  }
-  const onDelete = () => {
-    setData(null);
-  }
-  const onNext = () => {
-    if (data.steps < maxSteps) data.steps++;
-    setData({...data});
-  }
-  const onPrev = () => {
-    if (data.steps > 0) data.steps--;
-    setData({...data});
-  }
-  const onCollapse = () => {
-    setData({...data, collapsed: !data.collapsed});
-  }
-  return {
-    onEdit: onEdit,
-    onDelete: onDelete,
-    onNext: onNext,
-    onPrev: onPrev,
-    onCollapse: onCollapse,
-  }
 }
 
-export { getSchemaTable, getSavedPremiseTables, getTableButtonHandlers };
+export {
+  computeTable,
+  getTableDimensions, 
+  getLegend,
+  getTableModel,
+  getTableButtonHandlers,
+};
