+++ src/Display/EditorTable/EditorTable.js
@@ -1,19 +1,17 @@
 import React from 'react';
-import { getSchemaTable } from '../displayTableLogic';
+import SchemaTable from '../SavedTables/SchemaTable/SchemaTable';
 
 const EditorTable = (props) => {
   const { sentenceLetters, schema, setSchema } = props;
-  const editorTableData = {
-    sentenceLetters: sentenceLetters,
-    schema: schema,
-    showButtons: false,
-    clickHandlers : {
-      onSwitch: () => setSchema({...schema, type: schema.type === 'P' ? 'C' : 'P'})
-    }
-  }
-  const editorTable = getSchemaTable(editorTableData);
   return (
-    <div>{ editorTable }</div>
+    <SchemaTable
+      schema={schema}
+      sentenceLetters={sentenceLetters}
+      isSavedTable={false}
+      clickHandlers={{
+        onSwitch: () => setSchema({...schema, type: schema.type === 'P' ? 'C' : 'P'})
+      }} 
+    />
   );
 }
